+++ src/mlpack/core.hpp
@@ -59,6 +59,7 @@
  * - mlpack_approx_kfn
  * - mlpack_cf
  * - mlpack_decision_stump
+ * - mlpack_decision_tree
  * - mlpack_det
  * - mlpack_emst
  * - mlpack_fastmks
@@ -203,6 +204,14 @@
  *   - Anuraj Kanodia <akanuraj200@gmail.com>
  *   - Ivan Georgiev <ivan@jonan.info>
  *   - Shikhar Bhardwaj <shikharbhardwaj68@gmail.com>
+ *   - Yashu Seth <yashuseth2503@gmail.com>
+ *   - Mike Izbicki <mike@izbicki.me>
+ *   - Sudhanshu Ranjan <sranjan.sud@gmail.com>
+ *   - Piyush Jaiswal <piyush.jaiswal@st.niituniversity.in>
+ *   - Dinesh Raj <dinu.iota@gmail.com>
+ *   - Lakshya Agrawal <zeeshan.lakshya@gmail.com>
+ *   - Vivek Pal <vivekpal.dtu@gmail.com>
+ *   - Praveen Ch <chvsp972911@gmail.com>
  */
 
 // First, include all of the prerequisites.