+++ src/mlpack/methods/decision_stump/decision_stump.hpp
@@ -12,7 +12,7 @@
 #ifndef MLPACK_METHODS_DECISION_STUMP_DECISION_STUMP_HPP
 #define MLPACK_METHODS_DECISION_STUMP_DECISION_STUMP_HPP
 
-#include <mlpack/core.hpp>
+#include <mlpack/prereqs.hpp>
 
 namespace mlpack {
 namespace decision_stump {
@@ -83,6 +83,23 @@ class DecisionStump
    */
   void Train(const MatType& data,
              const arma::Row<size_t>& labels,
+             const size_t classes,
+             const size_t bucketSize);
+
+  /**
+   * Train the decision stump on the given data, with the given weights.  This
+   * completely overwrites any previous training data, so after training the
+   * stump may be completely different.
+   *
+   * @param data Dataset to train on.
+   * @param labels Labels for each point in the dataset.
+   * @param weights Weights for each point in the dataset.
+   * @param classes Number of classes in the dataset.
+   * @param bucketSize Minimum size of bucket when splitting.
+   */
+  void Train(const MatType& data,
+             const arma::Row<size_t>& labels,
+             const arma::rowvec& weights,
              const size_t classes,
              const size_t bucketSize);