+++ src/mlpack/methods/naive_bayes/nbc_main.cpp
@@ -12,7 +12,11 @@
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#include <mlpack/core.hpp>
+#include <mlpack/prereqs.hpp>
+#include <mlpack/core/util/param.hpp>
+#include <mlpack/core/data/normalize_labels.hpp>
+#include <mlpack/core/data/load.hpp>
+#include <mlpack/core/data/save.hpp>
 
 #include "naive_bayes_classifier.hpp"
 
@@ -21,7 +25,7 @@ PROGRAM_INFO("Parametric Naive Bayes Classifier",
     "training set, or loads a model from the given model file, and then may use"
     " that trained model to classify the points in a given test set."
     "\n\n"
-    "Labels are expected to be the last row of the training set (--train_file),"
+    "Labels are expected to be the last row of the training set (--training_file),"
     " but labels can also be passed in separately as their own file "
     "(--labels_file).  If training is not desired, a pre-existing model can be "
     "loaded with the --input_model_file (-m) option."