+++ src/mlpack/tests/distribution_test.cpp
@@ -20,6 +20,7 @@
 
 using namespace mlpack;
 using namespace mlpack::distribution;
+using namespace mlpack::math;
 
 BOOST_AUTO_TEST_SUITE(DistributionTest);
 
@@ -63,18 +64,18 @@ BOOST_AUTO_TEST_CASE(DiscreteDistributionProbabilityTest)
  */
 BOOST_AUTO_TEST_CASE(DiscreteDistributionRandomTest)
 {
-  DiscreteDistribution d(3);
+  DiscreteDistribution d(arma::Col<size_t>("3"));
 
   d.Probabilities() = "0.3 0.6 0.1";
 
   arma::vec actualProb(3);
+
   actualProb.zeros();
 
   for (size_t i = 0; i < 50000; i++)
     actualProb((size_t) (d.Random()[0] + 0.5))++;
 
   // Normalize.
-  Log::Debug << actualProb.t();
   actualProb /= accu(actualProb);
 
   // 8% tolerance, because this can be a noisy process.
@@ -116,6 +117,75 @@ BOOST_AUTO_TEST_CASE(DiscreteDistributionTrainProbTest)
   BOOST_REQUIRE_CLOSE(d.Probability("0"), 0.25, 1e-5);
   BOOST_REQUIRE_CLOSE(d.Probability("1"), 0.25, 1e-5);
   BOOST_REQUIRE_CLOSE(d.Probability("2"), 0.5, 1e-5);
+}
+
+/**
+ * Achieve multidimensional probability distribution.
+ */
+BOOST_AUTO_TEST_CASE(MultiDiscreteDistributionTrainProbTest)
+{
+  DiscreteDistribution d("10 10 10");
+
+  arma::mat obs("0 1 1 1 2 2 2 2 2 2;"
+                "0 0 0 1 1 1 2 2 2 2;"
+                "0 0 0 1 1 2 2 2 2 2;");
+ 
+  d.Train(obs);
+  BOOST_REQUIRE_CLOSE(d.Probability("0 0 0"), 0.009, 1e-5);
+  BOOST_REQUIRE_CLOSE(d.Probability("0 1 2"), 0.015, 1e-5);
+  BOOST_REQUIRE_CLOSE(d.Probability("2 1 0"), 0.054, 1e-5);
+}
+
+/**
+ * Make sure we initialize multidimensional probability distribution 
+ * correctly.
+ */
+BOOST_AUTO_TEST_CASE(MultiDiscreteDistributionConstructorTest)
+{
+  DiscreteDistribution d("4 4 4 4");
+
+  BOOST_REQUIRE_EQUAL(d.Probabilities(0).size(), 4);
+  BOOST_REQUIRE_EQUAL(d.Dimensionality(), 4);
+  BOOST_REQUIRE_CLOSE(d.Probability("0 0 0 0"), 0.00390625, 1e-5);
+  BOOST_REQUIRE_CLOSE(d.Probability("0 1 2 3"), 0.00390625, 1e-5);
+}
+
+/**
+ * Achieve multidimensional probability distribution.
+ */
+BOOST_AUTO_TEST_CASE(MultiDiscreteDistributionTrainTest)
+{
+  std::vector<arma::vec> pro;
+  pro.push_back(arma::vec("0.1, 0.3, 0.6"));
+  pro.push_back(arma::vec("0.3, 0.3, 0.3"));
+  pro.push_back(arma::vec("0.25, 0.25, 0.5"));
+
+  DiscreteDistribution d(pro);
+
+  BOOST_REQUIRE_CLOSE(d.Probability("0 0 0"), 0.0083333, 1e-3);
+  BOOST_REQUIRE_CLOSE(d.Probability("0 1 2"), 0.0166666, 1e-3);
+  BOOST_REQUIRE_CLOSE(d.Probability("2 1 0"), 0.05, 1e-5);
+}
+
+/**
+ * Estimate multidimensional probability distribution from observations with
+ * probabilities.
+ */
+BOOST_AUTO_TEST_CASE(MultiDiscreteDistributionTrainProTest)
+{
+  DiscreteDistribution d("5 5 5");
+
+  arma::mat obs("0 0 1 1 2;"
+                "0 1 1 2 2;"
+                "0 1 1 2 2");
+
+  arma::vec prob("0.25 0.25 0.25 0.25 1");
+
+  d.Train(obs, prob);
+
+  BOOST_REQUIRE_CLOSE(d.Probability("0 0 0"), 0.00390625, 1e-5);
+  BOOST_REQUIRE_CLOSE(d.Probability("1 0 1"), 0.0078125, 1e-5);
+  BOOST_REQUIRE_CLOSE(d.Probability("2 1 0"), 0.015625, 1e-5);
 }
 
 /*********************************/
@@ -390,10 +460,145 @@ BOOST_AUTO_TEST_CASE(GaussianDistributionTrainTest)
       BOOST_REQUIRE_SMALL(d.Covariance()(i, j) - actualCov(i, j), 1e-5);
 }
 
+/**
+  * This test verifies the fitting of GaussianDistribution
+  works properly when probabilities for each sample is given.
+**/
+BOOST_AUTO_TEST_CASE(GaussianDistributionTrainWithProbabilitiesTest)
+{
+  double mean = 5.0;
+  double stddeviation = 2.0;
+
+  //Creates a normal distribution generator.
+  std::default_random_engine generator;
+  generator.seed(std::time(NULL));
+  std::normal_distribution<double> dist(mean, stddeviation);
+  
+  size_t N = 50000;
+  size_t d =1;
+  
+  arma::mat rdata(d, N);
+  
+  for(size_t i = 0;i < d; i++)
+    for(size_t j = 0;j < N;j++)
+      rdata(i,j) = dist(generator);
+
+  //Creates a uniform distribution generator
+  std::uniform_real_distribution<double> prob(0, 1);
+  arma::vec probabilities(N);
+  
+  for(size_t i = 0;i < N;i++)
+    probabilities(i) = prob(generator);
+  
+  //Fits result with probabilities and data.
+  GaussianDistribution guDist;
+  guDist.Train(rdata, probabilities);
+  
+  //Fits result only with data
+  GaussianDistribution guDist2;
+  guDist2.Train(rdata);
+  
+  BOOST_REQUIRE_CLOSE(guDist.Mean()[0], guDist2.Mean()[0], 5);
+  BOOST_REQUIRE_CLOSE(guDist.Covariance()[0], guDist2.Covariance()[0], 5);
+  
+  BOOST_REQUIRE_CLOSE(guDist.Mean()[0], mean, 5);
+  BOOST_REQUIRE_CLOSE(guDist.Covariance()[0], stddeviation*stddeviation, 5);
+}
+/**
+  *This test ensures that the same result is obtained when trained
+  with probabilities all set to 1 and with no probabilities at all
+**/
+BOOST_AUTO_TEST_CASE(GaussianDistributionWithProbabilties1Test)
+{
+  double mean = 5.0;
+  double stddeviation = 4.0;
+
+  //Create a normal distribution random generator
+  std::default_random_engine generator;
+  generator.seed(std::time(NULL));
+  std::normal_distribution<double> dist(mean, stddeviation);
+
+  size_t N = 50000;
+  size_t d = 1;
+
+  arma::mat rdata(d, N);
+
+  for(size_t i = 0; i < d; i++)
+    for(size_t j = 0; j < N ; j++)
+      rdata(i,j) = dist(generator);
+
+  arma::vec probabilities(N, arma::fill::ones);
+
+  //fits data with only data
+  GaussianDistribution guDist;
+  guDist.Train(rdata);
+
+  //fits result with data and each probability as 1
+  GaussianDistribution guDist2;
+  guDist2.Train(rdata, probabilities);
+
+  BOOST_REQUIRE_CLOSE(guDist.Mean()[0], guDist2.Mean()[0], 1e-15);
+  BOOST_REQUIRE_CLOSE(guDist.Covariance()[0], guDist2.Covariance()[0], 1e-2);
+}
+/** This test draes points from two different normal distributions,
+  * stes the probabilities for points from the first distribution
+  * to something small and the probabilities for the second to
+  * something large
+  *It ensures that the normal distribution recovered the same
+  *parameters as the second normal distribution with high probabilities
+**/
+BOOST_AUTO_TEST_CASE(GaussianDistributionTrainWithTwoDistProbabilitiesTest)
+{
+  double mean1 = 5.0;
+  double stddeviation1 = 4.0;
+
+  double mean2 = 3.0;
+  double stddeviation2 = 1.0;
+
+  //Create two gaussian distribution random generator
+  std::default_random_engine generator;
+  generator.seed(std::time(NULL));
+  std::normal_distribution<double> dist1(mean1, stddeviation1);
+  std::normal_distribution<double> dist2(mean2, stddeviation2);
+
+  std::uniform_real_distribution<double> lowProb(0, 0.02);
+  std::uniform_real_distribution<double> highProb(0.98, 1);
+
+  size_t N = 50000;
+  size_t d = 1;
+
+  arma::mat rdata(d, N);
+  arma::vec probabilities(N);
+
+  //draws point alternatily from the two different distributions.
+  for(size_t i = 0 ; i < d; i++)
+  {
+    for(size_t j = 0; j < N; j++)
+    {
+      if(j%2 == 0)
+        rdata(i,j) = dist1(generator);
+      else
+        rdata(i,j) = dist2(generator);
+    }
+  }
+
+  for(size_t i = 0 ; i < N ; i++)
+  {
+    if(i%2 == 0)
+      probabilities(i) = highProb(generator);
+    else
+      probabilities(i) = lowProb(generator);
+  }
+
+  GaussianDistribution guDist;
+  guDist.Train(rdata, probabilities);
+
+  BOOST_REQUIRE_CLOSE(guDist.Mean()[0], mean1, 5);
+  BOOST_REQUIRE_CLOSE(guDist.Covariance()[0], stddeviation1*stddeviation1, 5);
+}
 /******************************/
 /** Gamma Distribution Tests **/
 /******************************/
-
 /**
  * Make sure that using an object to fit one reference set and then asking
  * to fit another works properly.
@@ -403,7 +608,6 @@ BOOST_AUTO_TEST_CASE(GammaDistributionTrainTest)
   // Create a gamma distribution random generator.
   double alphaReal = 5.3;
   double betaReal = 1.5;
-  std::default_random_engine generator;
   std::gamma_distribution<double> dist(alphaReal, betaReal);
 
   // Create a N x d gamma distribution data and fit the results.
@@ -414,7 +618,7 @@ BOOST_AUTO_TEST_CASE(GammaDistributionTrainTest)
   // Random generation of gamma-like points.
   for (size_t j = 0; j < d; ++j)
     for (size_t i = 0; i < N; ++i)
-      rdata(j, i) = dist(generator);
+      rdata(j, i) = dist(math::randGen);
 
   // Create Gamma object and call Train() on reference set.
   GammaDistribution gDist;
@@ -432,7 +636,7 @@ BOOST_AUTO_TEST_CASE(GammaDistributionTrainTest)
   // Random generation of gamma-like points.
   for (size_t j = 0; j < d2; ++j)
     for (size_t i = 0; i < N2; ++i)
-      rdata2(j, i) = dist(generator);
+      rdata2(j, i) = dist(math::randGen);
 
   // Fit results using old object.
   gDist.Train(rdata2);
@@ -440,6 +644,140 @@ BOOST_AUTO_TEST_CASE(GammaDistributionTrainTest)
   // Training must estimate d' pairs of alpha and beta parameters.
   BOOST_REQUIRE_EQUAL(gDist.Dimensionality(), d2);
   BOOST_REQUIRE_EQUAL(gDist.Dimensionality(), d2);
+}
+
+/**
+ * This test verifies that the fitting procedure for GammaDistribution works
+ * properly when probabilities for each sample is given.
+ */
+BOOST_AUTO_TEST_CASE(GammaDistributionTrainWithProbabilitiesTest)
+{
+  double alphaReal = 5.4;
+  double betaReal = 6.7;
+
+  // Create a gamma distribution random generator.
+  std::gamma_distribution<double> dist(alphaReal, betaReal);
+
+  size_t N = 50000;
+  size_t d = 2;
+  arma::mat rdata(d, N);
+
+  for (size_t j = 0; j < d; j++)
+    for (size_t i = 0; i < N; i++)
+      rdata(j, i) = dist(math::randGen);
+
+  // Fill the probabilities randomly.
+  arma::vec probabilities(N, arma::fill::randu);
+
+  // Fit results with probabilities and data.
+  GammaDistribution gDist;
+  gDist.Train(rdata, probabilities);
+
+  // Fit results with only data.
+  GammaDistribution gDist2;
+  gDist2.Train(rdata);
+
+  BOOST_REQUIRE_CLOSE(gDist2.Alpha(0), gDist.Alpha(0), 1.5);
+  BOOST_REQUIRE_CLOSE(gDist2.Beta(0), gDist.Beta(0), 1.5);
+
+  BOOST_REQUIRE_CLOSE(gDist2.Alpha(1), gDist.Alpha(1), 1.5);
+  BOOST_REQUIRE_CLOSE(gDist2.Beta(1), gDist.Beta(1), 1.5);
+
+  BOOST_REQUIRE_CLOSE(alphaReal, gDist.Alpha(0), 2.5);
+  BOOST_REQUIRE_CLOSE(betaReal, gDist.Beta(0), 2.5);
+
+  BOOST_REQUIRE_CLOSE(alphaReal, gDist.Alpha(1), 2.5);
+  BOOST_REQUIRE_CLOSE(betaReal, gDist.Beta(1), 2.5);
+}
+
+/**
+ * This test ensures that the same result is obtained when trained with
+ * probabilities all set to 1 and with no probabilities at all.
+ */
+BOOST_AUTO_TEST_CASE(GammaDistributionTrainAllProbabilities1Test)
+{
+  double alphaReal = 5.4;
+  double betaReal = 6.7;
+
+  // Create a gamma distribution random generator.
+  std::gamma_distribution<double> dist(alphaReal, betaReal);
+
+  size_t N = 1000;
+  size_t d = 2;
+  arma::mat rdata(d, N);
+
+  for (size_t j = 0; j < d; j++)
+    for (size_t i = 0; i < N; i++)
+      rdata(j, i) = dist(math::randGen);
+
+  // Fit results with only data.
+  GammaDistribution gDist;
+  gDist.Train(rdata);
+
+  // Fit results with data and each probability as 1.
+  GammaDistribution gDist2;
+  arma::vec allProbabilities1(N, arma::fill::ones);
+  gDist2.Train(rdata, allProbabilities1);
+
+  BOOST_REQUIRE_CLOSE(gDist2.Alpha(0), gDist.Alpha(0), 1e-5);
+  BOOST_REQUIRE_CLOSE(gDist2.Beta(0), gDist.Beta(0), 1e-5);
+
+  BOOST_REQUIRE_CLOSE(gDist2.Alpha(1), gDist.Alpha(1), 1e-5);
+  BOOST_REQUIRE_CLOSE(gDist2.Beta(1), gDist.Beta(1), 1e-5);
+}
+
+/**
+ * This test draws points from two different gamma distributions, sets the
+ * probabilities for the points from the first distribution to something small
+ * and the probabilities for the second to something large.  It ensures that the
+ * gamma distribution recovered has the same parameters as the second gamma
+ * distribution with high probabilities.
+ */
+BOOST_AUTO_TEST_CASE(GammaDistributionTrainTwoDistProbabilities1Test)
+{
+  double alphaReal = 5.4;
+  double betaReal = 6.7;
+
+  double alphaReal2 = 1.9;
+  double betaReal2 = 8.4;
+
+  // Create two gamma distribution random generators.
+  std::gamma_distribution<double> dist(alphaReal, betaReal);
+  std::gamma_distribution<double> dist2(alphaReal2, betaReal2);
+
+  size_t N = 50000;
+  size_t d = 2;
+  arma::mat rdata(d, N);
+  arma::vec probabilities(N);
+
+  // Draw points alternately from the two different distributions.
+  for (size_t j = 0; j < d; j++)
+  {
+    for (size_t i = 0; i < N; i++)
+    {
+      if (i % 2 == 0)
+        rdata(j, i) = dist(math::randGen);
+      else
+        rdata(j, i) = dist2(math::randGen);
+    }
+  }
+
+  for (size_t i = 0; i < N; i++)
+  {
+    if (i % 2 == 0)
+      probabilities(i) = 0.02 * math::Random();
+    else
+      probabilities(i) = 0.98 + 0.02 * math::Random();
+  }
+
+  GammaDistribution gDist;
+  gDist.Train(rdata, probabilities);
+
+  BOOST_REQUIRE_CLOSE(alphaReal2, gDist.Alpha(0), 5);
+  BOOST_REQUIRE_CLOSE(betaReal2, gDist.Beta(0), 5);
+
+  BOOST_REQUIRE_CLOSE(alphaReal2, gDist.Alpha(1), 5);
+  BOOST_REQUIRE_CLOSE(betaReal2, gDist.Beta(1), 5);
 }
 
 /**
@@ -455,22 +793,21 @@ BOOST_AUTO_TEST_CASE(GammaDistributionFittingTest)
   // better by a similar distribution.
   double errorTolerance = 10;
 
-  size_t N = 500;
+  size_t N = 5000;
   size_t d = 1; // Only 1 dimension is required for this.
 
   /** Iteration 1 (first parameter set) **/
 
-  // Create a gamma-random generator and data
+  // Create a gamma-random generator and data.
   double alphaReal = 5.3;
   double betaReal = 1.5;
-  std::default_random_engine generator;
   std::gamma_distribution<double> dist(alphaReal, betaReal);
 
   // Random generation of gamma-like points.
   arma::mat rdata(d, N);
   for (size_t j = 0; j < d; ++j)
     for (size_t i = 0; i < N; ++i)
-      rdata(j, i) = dist(generator);
+      rdata(j, i) = dist(math::randGen);
 
   // Create Gamma object and call Train() on reference set.
   GammaDistribution gDist;
@@ -482,17 +819,16 @@ BOOST_AUTO_TEST_CASE(GammaDistributionFittingTest)
 
   /** Iteration 2 (different parameter set) **/
 
-  // Create a gamma-random generator and data
+  // Create a gamma-random generator and data.
   double alphaReal2 = 7.2;
   double betaReal2 = 0.9;
-  std::default_random_engine generator2;
   std::gamma_distribution<double> dist2(alphaReal2, betaReal2);
 
   // Random generation of gamma-like points.
   arma::mat rdata2(d, N);
   for (size_t j = 0; j < d; ++j)
     for (size_t i = 0; i < N; ++i)
-      rdata2(j, i) = dist2(generator2);
+      rdata2(j, i) = dist2(math::randGen);
 
   // Create Gamma object and call Train() on reference set.
   GammaDistribution gDist2;
@@ -558,9 +894,8 @@ BOOST_AUTO_TEST_CASE(GammaDistributionRandomTest)
   arma::mat data(3, numPoints); // 3-d points.
 
   for (size_t i = 0; i < numPoints; ++i)
-    //std::cout << d1.Random() << "====" << std::endl;
     data.col(i) = d1.Random();
-  
+
   // Distribution to fit points.
   GammaDistribution d2(data);
   for (size_t i = 0; i < 3; ++i)
@@ -594,9 +929,8 @@ BOOST_AUTO_TEST_CASE(GammaDistributionProbabilityTest)
   // Combine into one 2-dimensional distribution.
   const arma::vec a3("2.0 3.1"), b3("0.9 1.4");
   arma::mat x3(2, 2);
-  x3
-    << 2.0 << 2.94 << arma::endr
-    << 2.0 << 2.94;
+  x3 << 2.0 << 2.94 << arma::endr
+     << 2.0 << 2.94;
   arma::vec prob3;
 
   // Expect that the 2-dimensional distribution returns the product of the