+++ src/mlpack/tests/hoeffding_tree_test.cpp
@@ -325,7 +325,7 @@ BOOST_AUTO_TEST_CASE(HoeffdingCategoricalSplitSplitTest)
 
   // No training is necessary because we can just call CreateChildren().
   data::DatasetInfo info(3);
-  info.MapString("hello", 0); // Make dimension 0 categorical.
+  info.MapString<size_t>("hello", 0); // Make dimension 0 categorical.
   HoeffdingCategoricalSplit<GiniImpurity>::SplitInfo splitInfo(3);
 
   // Create the children.
@@ -346,15 +346,15 @@ BOOST_AUTO_TEST_CASE(HoeffdingTreeNoSplitTest)
 {
   // Make all dimensions categorical.
   data::DatasetInfo info(3);
-  info.MapString("cat1", 0);
-  info.MapString("cat2", 0);
-  info.MapString("cat3", 0);
-  info.MapString("cat4", 0);
-  info.MapString("cat1", 1);
-  info.MapString("cat2", 1);
-  info.MapString("cat3", 1);
-  info.MapString("cat1", 2);
-  info.MapString("cat2", 2);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat2", 0);
+  info.MapString<size_t>("cat3", 0);
+  info.MapString<size_t>("cat4", 0);
+  info.MapString<size_t>("cat1", 1);
+  info.MapString<size_t>("cat2", 1);
+  info.MapString<size_t>("cat3", 1);
+  info.MapString<size_t>("cat1", 2);
+  info.MapString<size_t>("cat2", 2);
 
   HoeffdingTree<> split(info, 2, 0.95, 5000, 1);
 
@@ -383,9 +383,9 @@ BOOST_AUTO_TEST_CASE(HoeffdingTreeEasySplitTest)
   // will only receive points with class 1.  In the second dimension, all points
   // will have category 0 (so it is useless).
   data::DatasetInfo info(2);
-  info.MapString("cat0", 0);
-  info.MapString("cat1", 0);
-  info.MapString("cat0", 1);
+  info.MapString<size_t>("cat0", 0);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat0", 1);
 
   HoeffdingTree<> tree(info, 2, 0.95, 5000, 5000 /* never check for splits */);
 
@@ -411,9 +411,9 @@ BOOST_AUTO_TEST_CASE(HoeffdingTreeProbability1SplitTest)
   // will only receive points with class 1.  In the second dimension, all points
   // will have category 0 (so it is useless).
   data::DatasetInfo info(2);
-  info.MapString("cat0", 0);
-  info.MapString("cat1", 0);
-  info.MapString("cat0", 1);
+  info.MapString<size_t>("cat0", 0);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat0", 1);
 
   HoeffdingTree<> split(info, 2, 1.0, 12000, 1 /* always check for splits */);
 
@@ -438,10 +438,10 @@ BOOST_AUTO_TEST_CASE(HoeffdingTreeAlmostPerfectSplit)
 {
   // Two categories and two dimensions.
   data::DatasetInfo info(2);
-  info.MapString("cat0", 0);
-  info.MapString("cat1", 0);
-  info.MapString("cat0", 1);
-  info.MapString("cat1", 1);
+  info.MapString<size_t>("cat0", 0);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat0", 1);
+  info.MapString<size_t>("cat1", 1);
 
   HoeffdingTree<> split(info, 2, 0.95, 5000, 5000 /* never check for splits */);
 
@@ -473,10 +473,10 @@ BOOST_AUTO_TEST_CASE(HoeffdingTreeEqualSplitTest)
 {
   // Two categories and two dimensions.
   data::DatasetInfo info(2);
-  info.MapString("cat0", 0);
-  info.MapString("cat1", 0);
-  info.MapString("cat0", 1);
-  info.MapString("cat1", 1);
+  info.MapString<size_t>("cat0", 0);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat0", 1);
+  info.MapString<size_t>("cat1", 1);
 
   HoeffdingTree<> split(info, 2, 0.95, 5000, 1);
 
@@ -504,18 +504,18 @@ using HoeffdingSizeTNumericSplit = HoeffdingNumericSplit<FitnessFunction,
 BOOST_AUTO_TEST_CASE(HoeffdingTreeSimpleDatasetTest)
 {
   DatasetInfo info(3);
-  info.MapString("cat0", 0);
-  info.MapString("cat1", 0);
-  info.MapString("cat2", 0);
-  info.MapString("cat3", 0);
-  info.MapString("cat4", 0);
-  info.MapString("cat5", 0);
-  info.MapString("cat6", 0);
-  info.MapString("cat0", 1);
-  info.MapString("cat1", 1);
-  info.MapString("cat2", 1);
-  info.MapString("cat0", 2);
-  info.MapString("cat1", 2);
+  info.MapString<size_t>("cat0", 0);
+  info.MapString<size_t>("cat1", 0);
+  info.MapString<size_t>("cat2", 0);
+  info.MapString<size_t>("cat3", 0);
+  info.MapString<size_t>("cat4", 0);
+  info.MapString<size_t>("cat5", 0);
+  info.MapString<size_t>("cat6", 0);
+  info.MapString<size_t>("cat0", 1);
+  info.MapString<size_t>("cat1", 1);
+  info.MapString<size_t>("cat2", 1);
+  info.MapString<size_t>("cat0", 2);
+  info.MapString<size_t>("cat1", 2);
 
   // Now generate data.
   arma::Mat<size_t> dataset(3, 9000);
@@ -798,7 +798,7 @@ BOOST_AUTO_TEST_CASE(BinaryNumericHoeffdingTreeTest)
   arma::mat dataset(4, 9000);
   arma::Row<size_t> labels(9000);
   data::DatasetInfo info(4); // All features are numeric, except the fourth.
-  info.MapString("0", 3);
+  info.MapString<double>("0", 3);
   for (size_t i = 0; i < 9000; i += 3)
   {
     dataset(0, i) = mlpack::math::Random();
@@ -984,7 +984,7 @@ BOOST_AUTO_TEST_CASE(ConfidenceChangeTest)
   arma::mat dataset(4, 9000);
   arma::Row<size_t> labels(9000);
   data::DatasetInfo info(4); // All features are numeric, except the fourth.
-  info.MapString("0", 3);
+  info.MapString<double>("0", 3);
   for (size_t i = 0; i < 9000; i += 3)
   {
     dataset(0, i) = mlpack::math::Random();
@@ -1041,7 +1041,7 @@ BOOST_AUTO_TEST_CASE(ParameterChangeTest)
   arma::mat dataset(4, 9000);
   arma::Row<size_t> labels(9000);
   data::DatasetInfo info(4); // All features are numeric, except the fourth.
-  info.MapString("0", 3);
+  info.MapString<double>("0", 3);
   for (size_t i = 0; i < 9000; i += 3)
   {
     dataset(0, i) = mlpack::math::Random();
@@ -1094,7 +1094,7 @@ BOOST_AUTO_TEST_CASE(MultipleSerializationTest)
   arma::mat dataset(4, 9000);
   arma::Row<size_t> labels(9000);
   data::DatasetInfo info(4); // All features are numeric, except the fourth.
-  info.MapString("0", 3);
+  info.MapString<double>("0", 3);
   for (size_t i = 0; i < 9000; i += 3)
   {
     dataset(0, i) = mlpack::math::Random();