+++ src/mlpack/tests/load_save_test.cpp
@@ -12,6 +12,8 @@
 #include <sstream>
 
 #include <mlpack/core.hpp>
+#include <mlpack/core/data/load_arff.hpp>
+#include <mlpack/core/data/load_impl.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include "test_tools.hpp"
@@ -28,9 +30,7 @@ BOOST_AUTO_TEST_SUITE(LoadSaveTest);
 BOOST_AUTO_TEST_CASE(NoExtensionLoad)
 {
   arma::mat out;
-  Log::Warn.ignoreInput = true;
   BOOST_REQUIRE(data::Load("noextension", out) == false);
-  Log::Warn.ignoreInput = false;
 }
 
 /**
@@ -39,9 +39,7 @@ BOOST_AUTO_TEST_CASE(NoExtensionLoad)
 BOOST_AUTO_TEST_CASE(NoExtensionSave)
 {
   arma::mat out;
-  Log::Warn.ignoreInput = true;
   BOOST_REQUIRE(data::Save("noextension", out) == false);
-  Log::Warn.ignoreInput = false;
 }
 
 /**
@@ -50,9 +48,7 @@ BOOST_AUTO_TEST_CASE(NoExtensionSave)
 BOOST_AUTO_TEST_CASE(NotExistLoad)
 {
   arma::mat out;
-  Log::Warn.ignoreInput = true;
   BOOST_REQUIRE(data::Load("nonexistentfile_______________.csv", out) == false);
-  Log::Warn.ignoreInput = false;
 }
 
 /**
@@ -184,6 +180,94 @@ BOOST_AUTO_TEST_CASE(LoadTransposedCSVTest)
   // Remove the file.
   remove("test_file.csv");
 }
+
+/**
+ * The test LoadColVecCSVTest, LoadMatinColVec, LoadRowVecCSVTest need to run in
+ * debug mode only; without debugging symbols, size checks are not performed and
+ * thus the exception will not be thrown.
+ */
+
+/**
+ * Make sure ColVec can be loaded.
+ */
+#ifdef DEBUG
+BOOST_AUTO_TEST_CASE(LoadColVecCSVTest)
+{
+  fstream f;
+  f.open("test_file.csv", fstream::out);
+
+  for (int i = 0; i < 8; ++i)
+    f << i << endl;
+
+  f.close();
+
+  arma::colvec test;
+  BOOST_REQUIRE(data::Load("test_file.csv", test, false) == true);
+
+  BOOST_REQUIRE_EQUAL(test.n_cols, 1);
+  BOOST_REQUIRE_EQUAL(test.n_rows, 8);
+
+  for (size_t i = 0; i < 8; ++i)
+    BOOST_REQUIRE_CLOSE(test[i], (double) (i), 1e-5);
+
+  // Remove the file.
+  remove("test_file.csv");
+}
+
+/**
+ * Make sure Load() throws an exception when trying to load a matrix into a
+ * colvec or rowvec.
+ */
+BOOST_AUTO_TEST_CASE(LoadMatinColVec)
+{
+  fstream f;
+  f.open("test_file.csv", fstream::out);
+
+  f << "1, 2" << endl;
+  f << "3, 4" << endl;
+
+  f.close();
+
+  arma::colvec coltest;
+  BOOST_REQUIRE_THROW(data::Load("test_file.csv", coltest, false),
+      std::logic_error);
+  Timer::Stop("loading_data");
+
+  arma::rowvec rowtest;
+  BOOST_REQUIRE_THROW(data::Load("test_file.csv", rowtest, false),
+      std::logic_error);
+  Timer::Stop("loading_data");
+
+  remove("test_file.csv");
+}
+
+/**
+ * Make sure that rowvecs can be loaded successfully.
+ */
+BOOST_AUTO_TEST_CASE(LoadRowVecCSVTest)
+{
+  fstream f;
+  f.open("test_file.csv", fstream::out);
+
+  for (int i = 0 ; i < 7; ++i)
+    f << i << ",";
+  f << "7";
+  f << endl;
+
+  f.close();
+
+  arma::rowvec test;
+  BOOST_REQUIRE(data::Load("test_file.csv", test, false) == true);
+
+  BOOST_REQUIRE_EQUAL(test.n_cols, 8);
+  BOOST_REQUIRE_EQUAL(test.n_rows, 1);
+
+  for (size_t i = 0; i < 8 ; ++i)
+    BOOST_REQUIRE_CLOSE(test[i], (double) (i) , 1e-5);
+
+  remove("test_file.csv");
+}
+#endif
 
 /**
  * Make sure TSVs can be loaded in transposed form.
@@ -457,9 +541,7 @@ BOOST_AUTO_TEST_CASE(LoadRawBinaryTest)
       == true);
 
   // Now reload through our interface.
-  Log::Warn.ignoreInput = true;
   BOOST_REQUIRE(data::Load("test_file.bin", test) == true);
-  Log::Warn.ignoreInput = false;
 
   BOOST_REQUIRE_EQUAL(test.n_rows, 1);
   BOOST_REQUIRE_EQUAL(test.n_cols, 8);
@@ -652,12 +734,10 @@ BOOST_AUTO_TEST_CASE(NoHDF5Test)
   test.randu(5, 5);
 
   // Stop warnings.
-  Log::Warn.ignoreInput = true;
   BOOST_REQUIRE(data::Save("test_file.h5", test) == false);
   BOOST_REQUIRE(data::Save("test_file.hdf5", test) == false);
   BOOST_REQUIRE(data::Save("test_file.hdf", test) == false);
   BOOST_REQUIRE(data::Save("test_file.he5", test) == false);
-  Log::Warn.ignoreInput = false;
 }
 #endif
 
@@ -837,9 +917,9 @@ BOOST_AUTO_TEST_CASE(DatasetInfoTest)
   }
 
   // Okay.  Add some mappings for dimension 3.
-  const size_t first = di.MapString("test_mapping_1", 3);
-  const size_t second = di.MapString("test_mapping_2", 3);
-  const size_t third = di.MapString("test_mapping_3", 3);
+  const size_t first = di.MapString<size_t>("test_mapping_1", 3);
+  const size_t second = di.MapString<size_t>("test_mapping_2", 3);
+  const size_t third = di.MapString<size_t>("test_mapping_3", 3);
 
   BOOST_REQUIRE_EQUAL(first, 0);
   BOOST_REQUIRE_EQUAL(second, 1);
@@ -999,10 +1079,10 @@ BOOST_AUTO_TEST_CASE(CategoricalCSVLoadTest00)
   BOOST_REQUIRE(info.Type(1) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(2) == Datatype::categorical);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("hello", 2), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("goodbye", 2), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("coffee", 2), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("confusion", 2), 3);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("hello", 2), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("goodbye", 2), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("coffee", 2), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("confusion", 2), 3);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 2), "hello");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 2), "goodbye");
@@ -1048,8 +1128,8 @@ BOOST_AUTO_TEST_CASE(CategoricalCSVLoadTest01)
   BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(3) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 0), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("", 0), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 0), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 0), 1);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 0), "1");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 0), "");
@@ -1092,8 +1172,8 @@ BOOST_AUTO_TEST_CASE(CategoricalCSVLoadTest02)
   BOOST_REQUIRE(info.Type(1) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("", 0), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 0), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 0), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 0), 0);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 0), "1");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 0), "");
@@ -1136,10 +1216,54 @@ BOOST_AUTO_TEST_CASE(CategoricalCSVLoadTest03)
   BOOST_REQUIRE(info.Type(1) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("", 0), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 0), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 0), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 0), 1);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 0), "");
+  BOOST_REQUIRE_EQUAL(info.UnmapString(1, 0), "1");
+
+  remove("test.csv");
+}
+
+BOOST_AUTO_TEST_CASE(CategoricalCSVLoadTest04)
+{
+  fstream f;
+  f.open("test.csv", fstream::out);
+  f << "200-DM, 1, 1" << endl;
+  f << "1, 1, 1" << endl;
+  f << "1, 1, 1" << endl;
+  f << "1, 1, 1" << endl;
+  f.close();
+
+  // Load the test CSV.
+  arma::umat matrix;
+  DatasetInfo info;
+  data::Load("test.csv", matrix, info, true);
+
+  BOOST_REQUIRE_EQUAL(matrix.n_cols, 4);
+  BOOST_REQUIRE_EQUAL(matrix.n_rows, 3);
+
+  BOOST_REQUIRE_EQUAL(matrix(0, 0), 0);
+  BOOST_REQUIRE_EQUAL(matrix(0, 1), 1);
+  BOOST_REQUIRE_EQUAL(matrix(0, 2), 1);
+  BOOST_REQUIRE_EQUAL(matrix(0, 3), 1);
+  BOOST_REQUIRE_EQUAL(matrix(1, 0), 1);
+  BOOST_REQUIRE_EQUAL(matrix(1, 1), 1);
+  BOOST_REQUIRE_EQUAL(matrix(1, 2), 1);
+  BOOST_REQUIRE_EQUAL(matrix(1, 3), 1);
+  BOOST_REQUIRE_EQUAL(matrix(2, 0), 1);
+  BOOST_REQUIRE_EQUAL(matrix(2, 1), 1);
+  BOOST_REQUIRE_EQUAL(matrix(2, 2), 1);
+  BOOST_REQUIRE_EQUAL(matrix(2, 3), 1);
+
+  BOOST_REQUIRE(info.Type(0) == Datatype::categorical);
+  BOOST_REQUIRE(info.Type(1) == Datatype::numeric);
+  BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
+
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("200-DM", 0), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 0), 1);
+
+  BOOST_REQUIRE_EQUAL(info.UnmapString(0, 0), "200-DM");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 0), "1");
 
   remove("test.csv");
@@ -1196,24 +1320,24 @@ BOOST_AUTO_TEST_CASE(CategoricalNontransposedCSVLoadTest00)
   BOOST_REQUIRE(info.Type(5) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(6) == Datatype::categorical);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 0), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("2", 0), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("hello", 0), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("3", 1), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("4", 1), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("goodbye", 1), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("5", 2), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("6", 2), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("coffee", 2), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("7", 3), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("8", 3), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("confusion", 3), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("9", 4), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("10", 4), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("hello", 4), 2);
-  BOOST_REQUIRE_EQUAL(info.MapString("13", 6), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("14", 6), 1);
-  BOOST_REQUIRE_EQUAL(info.MapString("confusion", 6), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 0), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("2", 0), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("hello", 0), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("3", 1), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("4", 1), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("goodbye", 1), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("5", 2), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("6", 2), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("coffee", 2), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("7", 3), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("8", 3), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("confusion", 3), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("9", 4), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("10", 4), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("hello", 4), 2);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("13", 6), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("14", 6), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("confusion", 6), 2);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 0), "1");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 0), "2");
@@ -1273,8 +1397,8 @@ BOOST_AUTO_TEST_CASE(CategoricalNontransposedCSVLoadTest01)
   BOOST_REQUIRE(info.Type(2) == Datatype::categorical);
   BOOST_REQUIRE(info.Type(3) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("", 2), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 2), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 2), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 2), 1);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 2), "");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 2), "1");
@@ -1318,8 +1442,8 @@ BOOST_AUTO_TEST_CASE(CategoricalNontransposedCSVLoadTest02)
   BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(3) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("", 1), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 1), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 1), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 1), 1);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 1), "");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 1), "1");
@@ -1363,13 +1487,58 @@ BOOST_AUTO_TEST_CASE(CategoricalNontransposedCSVLoadTest03)
   BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
   BOOST_REQUIRE(info.Type(3) == Datatype::numeric);
 
-  BOOST_REQUIRE_EQUAL(info.MapString("", 1), 0);
-  BOOST_REQUIRE_EQUAL(info.MapString("1", 1), 1);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("", 1), 0);
+  BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 1), 1);
 
   BOOST_REQUIRE_EQUAL(info.UnmapString(0, 1), "");
   BOOST_REQUIRE_EQUAL(info.UnmapString(1, 1), "1");
 
   remove("test.csv");
+}
+
+BOOST_AUTO_TEST_CASE(CategoricalNontransposedCSVLoadTest04)
+{
+    fstream f;
+    f.open("test.csv", fstream::out);
+    f << " 200-DM ,   1  , 1  " << endl;
+    f << "  1 , 1  , 1  " << endl;
+    f << "  1  ,   1  ,  1  " << endl;
+    f << "  1  , 1  , 1  " << endl;
+    f.close();
+
+    // Load the test CSV.
+    arma::umat matrix;
+    DatasetInfo info;
+    data::Load("test.csv", matrix, info, true, false); // No transpose.
+
+    BOOST_REQUIRE_EQUAL(matrix.n_cols, 3);
+    BOOST_REQUIRE_EQUAL(matrix.n_rows, 4);
+
+    BOOST_REQUIRE(info.Type(0) == Datatype::categorical);
+    BOOST_REQUIRE(info.Type(1) == Datatype::numeric);
+    BOOST_REQUIRE(info.Type(2) == Datatype::numeric);
+    BOOST_REQUIRE(info.Type(3) == Datatype::numeric);
+
+    BOOST_REQUIRE_EQUAL(matrix(0, 0), 0);
+    BOOST_REQUIRE_EQUAL(matrix(0, 1), 1);
+    BOOST_REQUIRE_EQUAL(matrix(0, 2), 1);
+    BOOST_REQUIRE_EQUAL(matrix(1, 0), 1);
+    BOOST_REQUIRE_EQUAL(matrix(1, 1), 1);
+    BOOST_REQUIRE_EQUAL(matrix(1, 2), 1);
+    BOOST_REQUIRE_EQUAL(matrix(2, 0), 1);
+    BOOST_REQUIRE_EQUAL(matrix(2, 1), 1);
+    BOOST_REQUIRE_EQUAL(matrix(2, 2), 1);
+    BOOST_REQUIRE_EQUAL(matrix(3, 0), 1);
+    BOOST_REQUIRE_EQUAL(matrix(3, 1), 1);
+    BOOST_REQUIRE_EQUAL(matrix(3, 2), 1);
+
+    BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("200-DM", 1), 0);
+    BOOST_REQUIRE_EQUAL(info.MapString<arma::uword>("1", 1), 1);
+
+    BOOST_REQUIRE_EQUAL(info.UnmapString(0, 1), "200-DM");
+    BOOST_REQUIRE_EQUAL(info.UnmapString(1, 1), "1");
+
+    remove("test.csv");
 }
 
 /**
@@ -1607,6 +1776,22 @@ BOOST_AUTO_TEST_CASE(BadDatasetInfoARFFTest)
       std::invalid_argument);
 
   remove("test.arff");
+}
+
+/**
+ * A test to check whether the arff loader is case insensitive to declarations:
+ * @relation, @attribute, @data.
+ */
+BOOST_AUTO_TEST_CASE(CaseTest)
+{
+  arma::mat dataset;
+
+  DatasetMapper<IncrementPolicy> info;
+
+  LoadARFF<double, IncrementPolicy>("casecheck.arff", dataset, info);
+
+  BOOST_CHECK_EQUAL(dataset.n_rows, 2);
+  BOOST_CHECK_EQUAL(dataset.n_cols, 3);
 }
 
 BOOST_AUTO_TEST_SUITE_END();