+++ src/mlpack/tests/mlpack_test.cpp
@@ -12,8 +12,6 @@
  */
 #define BOOST_TEST_MODULE mlpackTest
 
-#include <mlpack/core/util/log.hpp>
-
 #include <boost/version.hpp>
 
 // We only need to do this for old Boost versions.
@@ -23,6 +21,8 @@
 
 #include <boost/test/unit_test.hpp>
 #include "test_tools.hpp"
+
+#include <mlpack/prereqs.hpp>
 
 /**
  * Provide a global fixture for each test.